requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.53.1
h11==0.14.0
idna==3.7
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
pure_eval==0.2.3
pycparser==2.22
pyparsing==3.1.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308
pyzmq==26.2.0
requests==2.31.0
retrying==1.3.4
selenium==4.22.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
via==0.1.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
wordcloud==1.9.4
wsproto==1.2.0
zipp==3.20.2